+++ ext/krypt/core/krypt_io_out_bytes.c
@@ -20,7 +20,7 @@ typedef struct int_outstream_bytes_st {
 #define int_safe_cast(out, in)		krypt_safe_cast_outstream((out), (in), OUTSTREAM_TYPE_BYTES, int_outstream_bytes)
 
 static int_outstream_bytes* int_bytes_alloc(void);
-static int int_bytes_write(krypt_outstream *out, unsigned char *buf, int len);
+static size_t int_bytes_write(krypt_outstream *out, unsigned char *buf, size_t len);
 static void int_bytes_free(krypt_outstream *out);
 
 static krypt_outstream_interface interface_bytes = {
@@ -65,8 +65,8 @@ int_bytes_alloc(void)
     return ret;
 }
 
-static int
-int_bytes_write(krypt_outstream *outstream, unsigned char *buf, int len)
+static size_t
+int_bytes_write(krypt_outstream *outstream, unsigned char *buf, size_t len)
 {
     int_outstream_bytes *out;
